fix: replace ExtensionValue with RawValue
Both these enums have similar definitions and serve similar purposes. To
minimize duplication, remove ExtensionValue and replace it with RawValue
inside Extensions implementation. Additional functionality of
ExtensionValue is added to RawValue, and ExtensionKind is renamed to
RawValueKind.

This also resolves issues with serialization of extensions, as
ExtensionValue serialzation for maps and sequences was not implemented
correctly, unlike in RawValue.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>